Find starting values for predicted values plots from the data in the model frame, if not supplied

Find starting values for predicted values plots from the data in the model frame, if not supplied

Usage

get.value(var, prediction.values, model)

Arguments

var

the variable to establish a prediction value for

prediction.values

(possible) user-supplied list of values for variables

model

the model containing the dataframe for mean estimation if values are not user-supplied (and warn the user if we're taking the mean)

Author

Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan