Do consistent dummy checks for functions (GDTE/GDRF) that use a GECM model

Do consistent dummy checks for functions (GDTE/GDRF) that use a GECM model

Usage

what.to.return(
  return.plot,
  return.formulae,
  return.data,
  plot.out,
  dat.out,
  the.final.formulae
)

Arguments

return.plot

a TRUE/FALSE on whether the plot should be returned from the function

return.formulae

a TRUE/FALSE on whether the data from the effect should be returned from the function

Author

Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan